chore : added feature command make migration

+++ src/Controller/ArticleController.php
@@ -31,7 +31,7 @@ class ArticleController extends Controller
 
     public function edit(Request $request, array $args): Response
     {
-        $id = (int) ($args['id'] ?? 0);
+        $id = $args['id'] ?? 0;
         $item = $this->service->getById($id);
 
         if (!$item) {
@@ -52,7 +52,7 @@ class ArticleController extends Controller
 
     public function update(Request $request, array $args): Response
     {
-        $id = (int) ($args['id'] ?? 0);
+        $id = $args['id'] ?? 0;
         $data = $this->parseBody($request);
         $this->service->update($id, $data);
         
@@ -61,7 +61,7 @@ class ArticleController extends Controller
 
     public function destroy(Request $request, array $args): Response
     {
-        $id = (int) ($args['id'] ?? 0);
+        $id = $args['id'] ?? 0;
         $this->service->delete($id);
         
         return $this->factory->createResponse(200)->withHeader('HX-Refresh', 'true');
